Restructure the folders according to the Daniil's proposal. Move user repository implementation to the repository folder.
