+++ backend/src/main/kotlin/com/example/config/Monitoring.kt
@@ -1,4 +1,4 @@
-package com.example.infrastructure.config
+package com.example.config
 
 import io.ktor.server.application.*
 import io.ktor.server.plugins.calllogging.*
