+++ backend/src/main/kotlin/com/example/config/Security.kt
@@ -1,4 +1,4 @@
-package com.example.infrastructure.config
+package com.example.config
 
 import com.auth0.jwt.JWT
 import com.auth0.jwt.algorithms.Algorithm
