+++ backend/src/main/kotlin/com/example/model/ExposedUser.kt
@@ -1,4 +1,4 @@
-package com.example.domain.model
+package com.example.model
 
 import kotlinx.serialization.Serializable
