+++ backend/src/test/kotlin/com/example/ApplicationTest.kt
@@ -1,6 +1,6 @@
 package com.example
 
-import com.example.web.routes.configureRouting
+import com.example.routes.configureRouting
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import io.ktor.http.*
